+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/api/SettingsResources.scala
@@ -109,7 +109,16 @@ object TopTabType extends Enumeration {
     `type`: TopTabType.Value,
     url: String,
     requiredPermission: Option[String],
-    addAccessTokenInQueryParam: Option[Boolean]
+    // Deprecated: use accessTokenInQuery.enabled setting instead
+    addAccessTokenInQueryParam: Option[Boolean],
+    accessTokenInQuery: Option[AccessTokenInQueryTabSettings] = Some(AccessTokenInQueryTabSettings())
+)
+
+@JsonCodec final case class AccessTokenInQueryTabSettings(
+    enabled: Boolean = false,
+    // The default parameter name is consistent with parameter name used by Grafana:
+    // https://grafana.com/docs/grafana/latest/setup-grafana/configure-security/configure-authentication/jwt/?plcmt=learn-nav#url-login
+    parameterName: Option[String] = Some("auth_token")
 )
 
 @JsonCodec final case class ToggleFeaturesOptions(